Осуществил переезд на javax и добавил слой для контроллера
